fmtutil.GetTermSize moved to terminal.GetSize
