--- fmtutil/termsize_windows.go
@@ -1,15 +1,0 @@
-// +build windows
-
-package fmtutil
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-//                                                                                    //
-//                     Copyright (c) 2009-2016 Essential Kaos                         //
-//      Essential Kaos Open Source License <http://essentialkaos.com/ekol?en>         //
-//                                                                                    //
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-// GetTermSize return window width and height
-func GetTermSize() (int, int) {
-	return -1, -1
-}
